+++ src/Stryker.Core/Stryker.Core.UnitTest/Options/StrykerInputsTests.cs
@@ -45,7 +45,8 @@ namespace Stryker.Core.UnitTest.Options
             ThresholdBreakInput = new ThresholdBreakInput(),
             ThresholdHighInput = new ThresholdHighInput(),
             ThresholdLowInput = new ThresholdLowInput(),
-            WithBaselineInput = new WithBaselineInput()
+            WithBaselineInput = new WithBaselineInput(),
+            BreakOnInitialTestFailureInput = new BreakOnInitialTestFailureInput()
         };
 
         [Fact]